Move `Input` to its own `text::input` module
