+++ core/src/text/editor.rs
@@ -1,22 +1,15 @@
 //! Edit text.
-#![allow(missing_docs)]
-use crate::alignment;
 use crate::clipboard;
 use crate::input_method;
 use crate::keyboard;
 use crate::keyboard::key;
-use crate::layout;
 use crate::mouse;
 use crate::renderer;
 use crate::text::highlighter::{self, Highlighter};
-use crate::text::paragraph;
-use crate::text::{self, Alignment, LineHeight, Text, Wrapping};
+use crate::text::{self, Alignment, LineHeight, Wrapping};
 use crate::widget::operation::Focusable;
 use crate::window;
-use crate::{
-    Color, Event, InputMethod, Length, Padding, Pixels, Point, Rectangle, Shell, Size, SmolStr,
-    Vector,
-};
+use crate::{Color, Event, InputMethod, Padding, Pixels, Point, Rectangle, Size, SmolStr, Vector};
 
 use std::borrow::Cow;
 use std::sync::Arc;
@@ -77,6 +70,7 @@ pub trait Editor: Sized + Default {
         new_highlighter: &mut impl Highlighter,
     );
 
+    /// Overwrites the current contents of the [`Editor`].
     fn overwrite(&mut self, new_text: &str);
 
     /// Runs a text [`Highlighter`] in the [`Editor`].
@@ -115,8 +109,13 @@ pub trait Editor: Sized + Default {
         contents
     }
 
+    /// Returns the current [`Font`](Self::Font) of the [`Editor`].
     fn font(&self) -> Self::Font;
+
+    /// Returns the current text size of the [`Editor`].
     fn text_size(&self) -> Pixels;
+
+    /// Returns the current [`LineHeight`] of the [`Editor`].
     fn line_height(&self) -> LineHeight;
 }
 
@@ -312,7 +311,7 @@ impl LineEnding {
     }
 }
 
-/// The state of an [`Editor`].
+/// The internal state of an [`Editor`].
 #[derive(Debug, Clone, Default)]
 pub struct State {
     focus: Option<Focus>,
@@ -323,17 +322,19 @@ pub struct State {
 }
 
 impl State {
+    /// Creates a new [`State`].
     pub fn new() -> Self {
         Self::default()
     }
 
+    /// Updates the [`State`] for the given [`Editor`] and returns any relevant [`Update`].
     pub fn update<Message>(
         &mut self,
+        editor: &impl Editor,
         event: &Event,
         bounds: Rectangle,
         padding: Padding,
         cursor: mouse::Cursor,
-        editor: &impl Editor,
         key_binding: impl Fn(KeyPress) -> Option<Binding<Message>>,
     ) -> Option<Update<Message>> {
         match event {
@@ -562,6 +563,7 @@ impl State {
         }
     }
 
+    /// Returns the current [`InputMethod`] of the [`State`] for the given [`Editor`].
     pub fn input_method<'a>(
         &'a self,
         editor: &impl Editor,
@@ -593,10 +595,11 @@ impl State {
         }
     }
 
+    /// Draws the given [`Editor`] with the current [`State`].
     pub fn draw<Renderer: text::Renderer>(
         &self,
-        renderer: &mut Renderer,
         editor: &Renderer::Editor,
+        renderer: &mut Renderer,
         position: Point,
         clip_bounds: Rectangle,
         style: Style,
@@ -660,15 +663,19 @@ impl State {
         }
     }
 
+    /// Returns whether the cursor of the [`Editor`] is visible.
     pub fn is_cursor_visible(&self) -> bool {
         self.focus.as_ref().is_some_and(Focus::is_cursor_visible)
     }
 }
 
+/// The visual style of an [`Editor`].
 pub struct Style {
+    /// The [`Color`] of the contents.
     pub value: Color,
+
+    /// The background [`Color`] of any selection.
     pub selection: Color,
-    pub placeholder: Color,
 }
 
 #[derive(Debug, Clone)]
@@ -719,20 +726,20 @@ impl Focusable for State {
     }
 }
 
-/// A binding to an action in the [`TextEditor`].
+/// A binding to an action in the [`Editor`].
 #[derive(Debug, Clone, PartialEq)]
 pub enum Binding<Message> {
-    /// Unfocus the [`TextEditor`].
+    /// Unfocus the [`Editor`].
     Unfocus,
-    /// Copy the selection of the [`TextEditor`].
+    /// Copy the selection of the [`Editor`].
     Copy,
-    /// Cut the selection of the [`TextEditor`].
+    /// Cut the selection of the [`Editor`].
     Cut,
-    /// Paste the clipboard contents in the [`TextEditor`].
+    /// Paste the clipboard contents in the [`Editor`].
     Paste,
-    /// Undo the last change peformed in the [`TextEditor`].
+    /// Undo the last change peformed in the [`Editor`].
     Undo,
-    /// Redo the last change undone in the [`TextEditor`].
+    /// Redo the last change undone in the [`Editor`].
     Redo,
     /// Apply a [`Motion`].
     Move(Motion),
@@ -860,16 +867,27 @@ impl<Message> Binding<Message> {
     }
 }
 
+/// The update of an [`Editor`], returned by [`State::update`].
 pub enum Update<Message> {
+    /// An [`Action`] must be performed in the [`Editor`].
     Action(Action),
+    /// The [`Editor`] just gained focus.
     Focus,
+    /// The [`Editor`] just lost focus.
     Unfocus,
-    Release,
+    /// The [`Editor`] changed its [`InputMethod`].
     InputMethod,
+    /// A mouse press was released in the [`Editor`].
+    Release,
+    /// The [`Editor`] must copy some text to the clipboard.
     Copy(String),
+    /// The [`Editor`] must paste the clipboard contents.
     Paste,
+    /// The [`Editor`] must be redrawn at the given [`Instant`].
     RedrawAt(Instant),
+    /// The [`Editor`] produced a custom message that must be published.
     Custom(Message),
+    /// The [`Editor`] produced a sequence of updates.
     Sequence(Vec<Update<Message>>),
 }
 
@@ -907,234 +925,4 @@ fn convert_macos_shortcut(
     };
 
     Some(keyboard::Key::Named(key))
-}
-
-pub struct Input<R: text::Renderer> {
-    editor: R::Editor,
-    state: State,
-    placeholder: paragraph::Plain<R::Paragraph>,
-    padding: Padding,
-}
-
-pub struct Layout<'a, Font> {
-    pub width: Length,
-    pub height: Length,
-    pub padding: Padding,
-    pub placeholder: &'a str,
-    pub font: Option<Font>,
-    pub size: Option<Pixels>,
-    pub line_height: LineHeight,
-    pub alignment: Alignment,
-    pub wrapping: Wrapping,
-}
-
-impl<R: text::Renderer> Input<R> {
-    pub fn new() -> Self {
-        Self {
-            editor: R::Editor::with_text(""),
-            state: State::new(),
-            placeholder: paragraph::Plain::default(),
-            padding: Padding::default(),
-        }
-    }
-
-    pub fn is_empty(&self) -> bool {
-        self.editor.is_empty()
-    }
-
-    pub fn value(&self) -> String {
-        self.editor.text()
-    }
-
-    pub fn placeholder(&self) -> &str {
-        self.placeholder.content()
-    }
-
-    pub fn is_focused(&self) -> bool {
-        self.state.is_focused()
-    }
-
-    pub fn focus(&mut self) {
-        self.state.focus();
-    }
-
-    pub fn unfocus(&mut self) {
-        self.state.unfocus();
-    }
-
-    pub fn overwrite(&mut self, value: &str) {
-        self.editor.overwrite(value);
-    }
-
-    pub fn layout(
-        &mut self,
-        renderer: &R,
-        limits: &layout::Limits,
-        layout: Layout<'_, R::Font>,
-    ) -> layout::Node {
-        self.padding = layout.padding;
-
-        let limits = limits
-            .width(layout.width)
-            .height(layout.height)
-            .shrink(layout.padding);
-
-        let font = layout.font.unwrap_or_else(|| renderer.default_font());
-        let size = layout.size.unwrap_or_else(|| renderer.default_size());
-        let hint_factor = renderer.hint_factor();
-
-        self.editor.update(
-            limits.max(),
-            font,
-            size,
-            layout.line_height,
-            layout.wrapping,
-            layout.alignment,
-            hint_factor,
-            &mut highlighter::PlainText,
-        );
-
-        let bounds = match layout.height {
-            Length::Fill
-            | Length::FillPortion(_)
-            | Length::Fixed(_)
-            | Length::Bounded { .. }
-            | Length::Fluid(_) => limits.max(),
-            Length::Shrink | Length::Fit => {
-                limits.resolve(layout.width, layout.height, self.editor.min_bounds())
-            }
-        };
-
-        let _ = self.placeholder.update(Text {
-            content: layout.placeholder,
-            font,
-            line_height: layout.line_height,
-            bounds,
-            size,
-            align_x: layout.alignment,
-            align_y: alignment::Vertical::Top,
-            shaping: text::Shaping::Advanced,
-            wrapping: text::Wrapping::None,
-            ellipsis: text::Ellipsis::None,
-            hint_factor,
-        });
-
-        layout::Node::new(bounds.expand(layout.padding))
-    }
-
-    pub fn update<Message>(
-        &mut self,
-        event: &Event,
-        bounds: Rectangle,
-        cursor: mouse::Cursor,
-        shell: &mut Shell<'_, Message>,
-        key_binding: impl Fn(KeyPress) -> Option<Binding<Message>>,
-    ) -> bool {
-        fn apply<Message>(
-            editor: &mut impl Editor,
-            shell: &mut Shell<'_, Message>,
-            update: Update<Message>,
-        ) -> bool {
-            match update {
-                Update::Action(action) => {
-                    let is_edit = action.is_edit();
-
-                    editor.perform(action);
-
-                    shell.request_redraw();
-                    shell.capture_event();
-
-                    return is_edit;
-                }
-                Update::Focus | Update::InputMethod => {
-                    shell.request_redraw();
-                    shell.capture_event();
-                }
-                Update::Unfocus => {
-                    shell.request_redraw();
-                }
-                Update::Release => {}
-                Update::Copy(text) => {
-                    shell.write_clipboard(text);
-                    shell.capture_event();
-                }
-                Update::Paste => {
-                    shell.read_clipboard(clipboard::Kind::Text);
-                    shell.capture_event();
-                }
-                Update::RedrawAt(at) => {
-                    shell.request_redraw_at(at);
-                }
-                Update::Custom(message) => {
-                    shell.publish(message);
-                    shell.capture_event();
-                }
-                Update::Sequence(updates) => {
-                    let mut is_edit = false;
-
-                    for update in updates {
-                        is_edit |= apply(editor, shell, update);
-                    }
-
-                    return is_edit;
-                }
-            }
-
-            false
-        }
-
-        let Some(update) = self.state.update(
-            event,
-            bounds,
-            self.padding,
-            cursor,
-            &self.editor,
-            key_binding,
-        ) else {
-            return false;
-        };
-
-        apply(&mut self.editor, shell, update)
-    }
-
-    pub fn draw(&self, renderer: &mut R, bounds: Rectangle, viewport: Rectangle, style: Style) {
-        let text_bounds = bounds.shrink(self.padding);
-
-        let Some(clip_bounds) = text_bounds.intersection(&viewport) else {
-            return;
-        };
-
-        if self.editor.is_empty() {
-            let anchor = text_bounds.anchor(
-                self.placeholder.min_bounds(),
-                self.placeholder.align_x(),
-                self.placeholder.align_y(),
-            );
-
-            renderer.fill_paragraph(
-                self.placeholder.raw(),
-                anchor,
-                style.placeholder,
-                clip_bounds,
-            );
-        }
-
-        self.state.draw(
-            renderer,
-            &self.editor,
-            text_bounds.position(),
-            clip_bounds,
-            style,
-        );
-    }
-
-    pub fn input_method(&self, position: Point) -> InputMethod<&str> {
-        self.state.input_method(&self.editor, position)
-    }
-}
-
-impl<R: text::Renderer> Default for Input<R> {
-    fn default() -> Self {
-        Self::new()
-    }
 }
